fq_api_automation/requirements.txt
allure-pytest
pytest
selenium==3.141.0
pytest-xdist==1.32.0
pbr==5.4.5
aircv==1.4.6
opencv-python
helium==3.0.4
openpyxl==3.1.3
cachetools
tkinter-tooltip==2.1.0
psutil==5.9.5
demjson3==3.0.6
pytest-dependency==0.6.0
filelock==3.12.2

